requirements.txt
PyQt5
Pillow
pillow_heif
requests
pyexiftool